requirements.txt
streamlit==1.16.0
openai>=0.25.0
python-dotenv>=0.21.0